# Support `dual_gap_` on `ElasticNet` & `Lasso`