This is to improve sklearn compatibility. Both `ElasticNet` and `Lasso` in sklearn compute the dual gap as part of the coordinate descent solver, and store it as `dual_gap_` on the fit model.

For now, we compute the `dual_gap_` from the final fit state and store it the same. Two notes:

- Our CD solver doesn't use the dual gap as part of its stopping criteria, while sklearn's does. This means that we in practice stop fitting earlier in the optimization, resulting in a larger dual gap for the same tolerance. We could (and maybe should) update our solver to better match that of sklearn's.
- Our CD solver computes much of what's needed for calculating the dual gap as part of the fit. Doing this with cupy after the fit repeats work. However, for problems of a meaningful size computing the dual gap afterwards is small compared to the cost of the fit. I think doing the easier thing and leaving things in python makes sense for now.

Fixes #6467.