@jcrist 2025-05-15
Dask tests are failing due to #6737. Looks like there's 2 xpassing tests in the sklearn test suite. If we think this is still worth merging, I'll fix and push again.