@jcrist 2025-05-19
Closing this since we'll want to handle this in the solver itself. I've opened #6759 to track that.